Projects/Stock Screener/requirements.txt
alpaca-py>=0.18.0
pandas>=2.0.0
numpy>=1.24.0
streamlit>=1.30.0
streamlit-autorefresh>=0.0.1
yfinance>=0.2.36
python-dotenv>=1.0.0
pytz>=2023.3
plotly>=5.0.0
